updated mod structure
moved folders to client/common folder
