+++ src/main/java/com/turtysproductions/humlands/common/blocks/ScrapSandBlock.java
@@ -1,0 +1,11 @@
+package com.turtysproductions.humlands.common.blocks;
+
+import net.minecraft.block.FallingBlock;
+
+public class ScrapSandBlock extends FallingBlock{
+
+	public ScrapSandBlock(Properties properties) {
+		super(properties);
+	}
+
+}
