+++ src/main/java/com/turtysproductions/humlands/common/tab/HumlandsTab.java
@@ -1,6 +1,6 @@
-package com.turtysproductions.humlands.item;
+package com.turtysproductions.humlands.common.tab;
 
-import com.turtysproductions.humlands.init.ItemInit;
+import com.turtysproductions.humlands.common.init.ItemInit;
 
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
