+++ src/main/java/com/turtysproductions/humlands/common/world/dimension/HumlandsModDimension.java
@@ -1,4 +1,4 @@
-package com.turtysproductions.humlands.world.dimension;
+package com.turtysproductions.humlands.common.world.dimension;
 
 import java.util.function.BiFunction;
