--- src/main/java/com/turtysproductions/humlands/objects/blocks/ScrapSandBlock.java
@@ -1,11 +1,0 @@
-package com.turtysproductions.humlands.objects.blocks;
-
-import net.minecraft.block.FallingBlock;
-
-public class ScrapSandBlock extends FallingBlock{
-
-	public ScrapSandBlock(Properties properties) {
-		super(properties);
-	}
-
-}
